LPS-123218 Make OrderedServiceTrackerMap internal
