+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/OrderedServiceTrackerMap.java
@@ -12,19 +12,25 @@
  *
  */
 
-package com.liferay.saml.opensaml.integration;
+package com.liferay.saml.opensaml.integration.internal;
 
-import java.util.function.Supplier;
-
-import org.osgi.framework.BundleContext;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * @author Stian Sigvartsen
  */
-public interface OrderedServiceTrackerMapFactory {
+public interface OrderedServiceTrackerMap<T> {
 
-	public <T> OrderedServiceTrackerMap<T> create(
-		BundleContext bundleContext, Class<T> clazz, String propertyKey,
-		Supplier<T> defaultServiceSupplier);
+	public void close();
+
+	public List<Map.Entry<String, T>> getOrderedServices();
+
+	public List<String> getOrderedServicesKeys();
+
+	public T getService(String key);
+
+	public Set<String> getServicesKeys();
 
 }
