+++ modules/dxp/apps/saml/saml-opensaml-integration/src/main/java/com/liferay/saml/opensaml/integration/internal/field/expression/resolver/registry/UserFieldExpressionResolverRegistryImpl.java
@@ -14,10 +14,10 @@
 
 package com.liferay.saml.opensaml.integration.internal.field.expression.resolver.registry;
 
-import com.liferay.saml.opensaml.integration.OrderedServiceTrackerMap;
-import com.liferay.saml.opensaml.integration.OrderedServiceTrackerMapFactory;
 import com.liferay.saml.opensaml.integration.field.expression.resolver.UserFieldExpressionResolver;
 import com.liferay.saml.opensaml.integration.field.expression.resolver.registry.UserFieldExpressionResolverRegistry;
+import com.liferay.saml.opensaml.integration.internal.OrderedServiceTrackerMap;
+import com.liferay.saml.opensaml.integration.internal.OrderedServiceTrackerMapFactory;
 
 import java.util.List;
 import java.util.Map;
